Move dashboard_institutions to dashboard view.
- Move tests.
