+++ website/routes.py
@@ -34,7 +34,7 @@ from website.util import metrics
 from website.util import paths
 from website.util import sanitize
 from website import maintenance
-from website.models import Institution, Node
+from website.models import Institution
 from website import landing_pages as landing_page_views
 from website import views as website_views
 from website.citations import views as citation_views
@@ -57,20 +57,6 @@ def get_globals():
     user = _get_current_user()
 
     user_institutions = [{'id': inst._id, 'name': inst.name, 'logo_path': inst.logo_path_rounded_corners} for inst in user.affiliated_institutions] if user else []
-    all_institutions = [inst for inst in Institution.find().sort('name')]
-    dashboard_institutions = [
-        {'id': inst._id, 'name': inst.name, 'logo_path': inst.logo_path_rounded_corners}
-        for inst in all_institutions
-        if len(
-            Node.find_by_institutions(inst, query=(
-                Q('is_public', 'eq', True) &
-                Q('is_folder', 'ne', True) &
-                Q('is_deleted', 'ne', True) &
-                Q('parent_node', 'eq', None) &
-                Q('is_registration', 'eq', False)
-            ))
-        ) >= settings.INSTITUTION_DISPLAY_NODE_THRESHOLD
-    ]
     location = geolite2.lookup(request.remote_addr) if request.remote_addr else None
     if request.host_url != settings.DOMAIN:
         try:
@@ -93,7 +79,6 @@ def get_globals():
         'user_api_url': user.api_url if user else '',
         'user_entry_point': metrics.get_entry_point(user) if user else '',
         'user_institutions': user_institutions if user else None,
-        'dashboard_institutions': dashboard_institutions,
         'display_name': get_display_name(user.fullname) if user else '',
         'anon': {
             'continent': getattr(location, 'continent', None),
