+++ website/templates/home.mako
@@ -26,5 +26,10 @@
 
 <%def name="javascript_bottom()">
     ${parent.javascript_bottom()}
+    <script type="text/javascript">
+        window.contextVars = $.extend(true, {}, window.contextVars, {
+            dashboardInstitutions: ${ dashboard_institutions | sjson, n},
+        });
+    </script>
     <script src=${"/static/public/js/home-page.js" | webpack_asset}></script>
 </%def>
